+++ apps/indexer/config/prod/erigon.exs
@@ -1,5 +1,16 @@
 import Config
 
+basic_auth_user = System.get_env("ETHEREUM_JSONRPC_USER", "")
+basic_auth_pass = System.get_env("ETHEREUM_JSONRPC_PASSWORD", "")
+
+hackney_opts =
+  [pool: :ethereum_jsonrpc]
+  |> (&if(System.get_env("ETHEREUM_JSONRPC_HTTP_INSECURE", "") == "true", do: [:insecure] ++ &1, else: &1)).()
+  |> (&if(basic_auth_user != "" && basic_auth_pass != "",
+        do: [basic_auth: {basic_auth_user, basic_auth_pass}] ++ &1,
+        else: &1
+      )).()
+
 config :indexer,
   block_interval: :timer.seconds(5),
   json_rpc_named_arguments: [
@@ -16,7 +27,7 @@ config :indexer,
         trace_block: System.get_env("ETHEREUM_JSONRPC_TRACE_URL"),
         trace_replayTransaction: System.get_env("ETHEREUM_JSONRPC_TRACE_URL")
       ],
-      http_options: [recv_timeout: :timer.minutes(10), timeout: :timer.minutes(10), hackney: [pool: :ethereum_jsonrpc]]
+      http_options: [recv_timeout: :timer.minutes(10), timeout: :timer.minutes(10), hackney: hackney_opts]
     ],
     variant: EthereumJSONRPC.Erigon
   ],