+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features
 
+- [#6897](https://github.com/blockscout/blockscout/pull/6897) - Support basic auth in JSON RPC endpoint
+
 ### Fixes
 
 - [#6891](https://github.com/blockscout/blockscout/pull/6891) - Fix read contract for geth